Week2: Add explicit casting to byte

+++ Samples/Program.cs
@@ -38,7 +38,7 @@ namespace KMA.ProgrammingInCSharp2026.Samples
         {
             int x = 1000;
 
-            byte y = x;
+            byte y = (byte)x;
 
             Console.WriteLine(y);
         }
